add support for multiple files upload

+++ app/src/main/java/xyz/hisname/fireflyiii/workers/transaction/AttachmentWorker.kt
@@ -24,13 +24,17 @@ class AttachmentWorker(private val context: Context, workerParameters: WorkerPar
     private val channelIcon = R.drawable.ic_refresh
 
     companion object {
-        fun initWorker(fileUri: Uri?, transactionJournalId: Long, context: Context){
+        fun initWorker(fileUri: ArrayList<Uri>, transactionJournalId: Long, context: Context){
             val dataBuilder = Data.Builder()
+            val uriArray = arrayListOf<String>()
+            fileUri.forEach { uri ->
+                uriArray.add(uri.toString())
+            }
             val workBuilder = OneTimeWorkRequest
                     .Builder(AttachmentWorker::class.java)
                     .addTag("attachment_worker")
                     .setInputData(dataBuilder.putLong("transactionJournalId" ,transactionJournalId).build())
-                    .setInputData(dataBuilder.putString("fileUri", fileUri.toString()).build())
+                    .setInputData(dataBuilder.putStringArray("fileUri", uriArray.toTypedArray()).build())
                     .setConstraints(Constraints.Builder()
                             .setRequiredNetworkType(NetworkType.CONNECTED)
                             .build())
@@ -41,45 +45,48 @@ class AttachmentWorker(private val context: Context, workerParameters: WorkerPar
 
     override suspend fun doWork(): Result {
         val transactionJournalId = inputData.getLong("transactionJournalId", 0)
-        val fileUri = inputData.getString("fileUri")?.toUri() ?: Uri.EMPTY
+        val fileArray = inputData.getStringArray("fileUri")
         val service = genericService?.create(AttachmentService::class.java)
-        val fileName = FileUtils.getFileName(context, fileUri) ?: ""
-        val requestFile = RequestBody.create(MediaType.parse(FileUtils.getMimeType(context, fileUri) ?: ""), copyFile(context, fileUri, fileName))
-        try {
-            val storeAttachment = service?.storeAttachment(fileName, "TransactionJournal", transactionJournalId, fileName,
-                    "File uploaded by " + BuildConfig.APPLICATION_ID)
-            val responseBody = storeAttachment?.body()
-            if (responseBody != null && storeAttachment.code() == 200) {
-                val upload = service.uploadFile(responseBody.data.attachmentId, requestFile)
-                ("file://" + context.filesDir.path + "/" + fileName).toUri().toFile().delete()
-                if(upload.code() == 204) {
-                    context.displayNotification("File uploaded", channelName,
-                            Constants.TRANSACTION_CHANNEL, channelIcon)
-                    return Result.success()
-                } else {
-                    context.displayNotification(upload.message(), channelName,
-                            Constants.TRANSACTION_CHANNEL, channelIcon)
-                    return Result.failure()
+        if(!fileArray.isNullOrEmpty()){
+            fileArray.forEach { fileToUpload ->
+                val fileName = FileUtils.getFileName(context, fileToUpload.toUri()) ?: ""
+                val requestFile = RequestBody.create(MediaType.parse(FileUtils.getMimeType(context, fileToUpload.toUri()) ?: ""),
+                        copyFile(context, fileToUpload.toUri(), fileName))
+                try {
+                    val storeAttachment = service?.storeAttachment(fileName, "TransactionJournal", transactionJournalId, fileName,
+                            "File uploaded by " + BuildConfig.APPLICATION_ID)
+                    val responseBody = storeAttachment?.body()
+                    if (responseBody != null && storeAttachment.code() == 200) {
+                        val upload = service.uploadFile(responseBody.data.attachmentId, requestFile)
+                        ("file://" + context.filesDir.path + "/" + fileName).toUri().toFile().delete()
+                        if(upload.code() == 204) {
+                            context.displayNotification("File uploaded", channelName,
+                                    Constants.TRANSACTION_CHANNEL, channelIcon)
+                        } else {
+                            context.displayNotification(upload.message(), channelName,
+                                    Constants.TRANSACTION_CHANNEL, channelIcon)
+                        }
+                    } else {
+                        ("file://" + context.filesDir.path + "/" + fileName).toUri().toFile().delete()
+                        context.displayNotification(storeAttachment?.message() ?: "", channelName,
+                                Constants.TRANSACTION_CHANNEL, channelIcon)
+                    }
+                } catch (exception: Exception){
+                    ("file://" + context.filesDir.path + "/" + fileName).toUri().toFile().delete()
+                    val throwMessage = exception.message
+                    if(throwMessage != null && throwMessage.startsWith("Write error: ssl=")){
+                        context.displayNotification("Unable to add attachment to transaction as " +
+                                "there is a file limit on server side", channelName,
+                                Constants.TRANSACTION_CHANNEL, channelIcon)
+                    } else {
+                        context.displayNotification(exception.localizedMessage, channelName,
+                                Constants.TRANSACTION_CHANNEL, channelIcon)
+                    }
                 }
-            } else {
-                ("file://" + context.filesDir.path + "/" + fileName).toUri().toFile().delete()
-                context.displayNotification(storeAttachment?.message() ?: "", channelName,
-                        Constants.TRANSACTION_CHANNEL, channelIcon)
-                return Result.failure()
             }
-        } catch (exception: Exception){
-            ("file://" + context.filesDir.path + "/" + fileName).toUri().toFile().delete()
-            val throwMessage = exception.message
-            if(throwMessage != null && throwMessage.startsWith("Write error: ssl=")){
-                context.displayNotification("Unable to add attachment to transaction as " +
-                        "there is a file limit on server side", channelName,
-                        Constants.TRANSACTION_CHANNEL, channelIcon)
-            } else {
-                context.displayNotification(exception.localizedMessage, channelName,
-                        Constants.TRANSACTION_CHANNEL, channelIcon)
-            }
-            return Result.failure()
         }
+        // Will always return success
+        return Result.success()
     }
 
     private fun copyFile(context: Context, fileUri: Uri, name: String): File{
